+++ src/main/java/org/sosly/arcaneadditions/networking/messages/ClientMessageHandler.java
@@ -17,17 +17,17 @@ public class ClientMessageHandler {
     public static <T extends BaseMessage> boolean validateBasics(T message, NetworkEvent.Context ctx) {
         LogicalSide sideReceived = ctx.getDirection().getReceptionSide();
         ctx.setPacketHandled(true);
-        Logger var10000;
-        String var10001;
+        Logger logger;
+        String messageName;
         if (sideReceived != LogicalSide.CLIENT) {
-            var10000 = ArcaneAdditions.LOGGER;
-            var10001 = message.getClass().getName();
-            var10000.error(var10001 + " received on wrong side: " + sideReceived);
+            logger = ArcaneAdditions.LOGGER;
+            messageName = message.getClass().getName();
+            logger.error(messageName + " received on wrong side: " + sideReceived);
             return false;
         } else if (!message.isMessageValid()) {
-            var10000 = ArcaneAdditions.LOGGER;
-            var10001 = message.getClass().getName();
-            var10000.error(var10001 + " was invalid: " + message);
+            logger = ArcaneAdditions.LOGGER;
+            messageName = message.getClass().getName();
+            logger.error(messageName + " was invalid: " + message);
             return false;
         } else {
             return true;